refactor: extract welcome + autocomplete (#357)
Fourteenth slice. Two siblings of the chat-area / composer regions:

- draw_welcome      (~115 lines) -> ui/welcome.rs
- draw_autocomplete (~135 lines) -> ui/autocomplete.rs

Both moved with their imports (AutocompleteMode and COMMANDS now only
used by autocomplete.rs; dropped from ui/mod.rs). Also cleaned up four
orphan /// doc comments left over from earlier extractions and replaced
build_poll_display's misleading three-line comment header with a one-
liner that actually describes what it does.

ui/mod.rs: 2466 -> 2218 lines.
